Almost there. We have to fix every remaining error
The new message system should work, we have to
adapt the existing code for using it.

Here is the idea.
Three kind of message: Order, Support, Immediate.

Orders have their own serialId, Support messages
refers to an existing serialId.

Nodes sort messages, they expect a certain serialId
and only when completed the order the serialId is
increased. Support messages are read if the node
is executing the relative serialId order.

Immediate messages ignore all this rules and are
consumed as fast as possible.

To keep all nodes syncronized Orders are sent to
all nodes.